+++ app/list/components/FilterBar.vue
@@ -10,7 +10,7 @@
 </template>
 
 <script>
-    import { debounce } from 'lodash';
+    import debounce from 'lodash/debounce';
 
     export default {
         props: ['store'],