+++ classes/Controllers/AbstractController.php
@@ -23,6 +23,7 @@ use Grav\Framework\RequestHandler\Exception\NotFoundException;
 use Grav\Framework\RequestHandler\Exception\PageExpiredException;
 use Grav\Framework\Route\Route;
 use Grav\Plugin\FlexObjects\Flex;
+use Grav\Plugin\Form\Forms;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\RequestHandlerInterface;
@@ -183,14 +184,16 @@ abstract class AbstractController implements RequestHandlerInterface
         }
 
         $formName = $this->getPost('__form-name__');
-        $uniqueId = $this->getPost('__unique_form_id__') ?: $formName;
-
-        $form = $object->getForm($type ?? 'edit');
-        if ($uniqueId) {
-            $form->setUniqueId($uniqueId);
+        if ($formName) {
+            /** @var Forms $forms */
+            $forms = $this->getGrav()['forms'];
+            $form = $forms->getActiveForm();
+            if ($form->getName() == $formName && $form instanceof FlexForm && $form->getObject()->getFlexKey() === $object->getFlexKey()) {
+                return $form;
+            }
         }
 
-        return $form;
+        return $object->getForm($type ?? 'edit');
     }
 
     /**